third_party/blink/web_tests/platform/generic/http/tests/inspector-protocol/runtime/runtime-evaluate-silent-expected.txt

Tests that Runtime.evaluate logs usage to use counter correctly when muted/unmuted.

Running test: testUseCounterMuted
UseCounter should be silenced when muted
{
    id : <number>
    result : {
        result : {
            type : boolean
            value : false
        }
    }
    sessionId : <string>
}
UseCounter should be unsilenced when unmuted
{
    id : <number>
    result : {
        result : {
            type : boolean
            value : true
        }
    }
    sessionId : <string>
}
